fix: refine gradient scrim display logic
