+++ src/layouts/sidebar-sidecar/components/sidecar-scroll-container/index.tsx
@@ -1,60 +1,41 @@
-import { ReactNode, useEffect, useState } from 'react'
+import { ReactNode, useEffect, useRef, useState } from 'react'
 import classNames from 'classnames'
 import useWindowSize from 'hooks/use-window-size'
+import { getScrollData, ScrollData } from './utils/get-scroll-data'
 import s from './sidecar-scroll-container.module.css'
-
-/**
- * Scroll data keeps track of whether we're at the start or end
- * of scrolling the sidecar contents. This informs whether we display
- * gradient scrims, which provide a visual clue for scrollability.
- */
-interface ScrollData {
-	isScrollable: boolean
-	isAtStart?: boolean
-	isAtEnd?: boolean
-}
 
 /**
  * Renders a scrollable container intended for sidecar contents,
  * with a gradient scrim at the bottom of the container.
+ *
+ * Gradient scrims at the top and bottom of the scrollable container
+ * are conditionally displayed, when the container is scrollable:
+ * - If we're more than 0% scrolled, a gradient scrim at the top is shown
+ * - If we're not yet 100% scrolled, a gradient scrim at the bottom is shown
  */
 function SidecarScrollContainer({ children }: { children: ReactNode }) {
 	const windowSize = useWindowSize()
-	const [scrollData, setScrollData] = useState<ScrollData>({
-		isScrollable: false,
-	})
+	const scrollRef = useRef<HTMLDivElement>()
+	const [{ isScrollable, isAtStart, isAtEnd }, setScrollData] =
+		useState<ScrollData>({
+			isScrollable: false,
+		})
 
 	/**
-	 * Handle scroll.
-	 * Gradient scrims at the top and bottom of the scrollable container
-	 * are conditionally displayed, when the container is scrollable:
-	 * - If we're more than 0% scrolled, a gradient scrim at the top is shown
-	 * - If we're not yet 100% scrolled, a gradient scrim at the bottom is shown
+	 * Handle scroll, updating data for gradient scrims.
+	 * Note this fires only when the scrollRef container is scrolled.
 	 */
 	function handleScroll(e) {
-		const { scrollTop, scrollHeight, clientHeight } = e.target
-		const scrollMax = scrollHeight - clientHeight
-		const scrollPercent = Math.round((100 * scrollTop) / scrollMax)
-		const isValid = !Number.isNaN(scrollPercent)
-		if (isValid) {
-			const isAtStart = isScrollable && scrollPercent === 0
-			const isAtEnd = isScrollable && scrollPercent === 100
-			setScrollData({ isScrollable: true, isAtEnd, isAtStart })
-		}
+		setScrollData(getScrollData(e.target))
 	}
 
 	/**
-	 * When the window is resized, reset to { isScrollable: false }.
-	 * This ensures we don't show a gradient scrim when it's not needed.
-	 * Note: we could get more fancy here, as sometimes we'll hide the scrim
-	 * when in fact the area is still scrollable, but the effect is subtle
-	 * enough that this seems to work well enough.
+	 * Handle window resize, updating data for gradient scrims.
+	 * Note this fires on initial load, and subsequent viewport resizing.
 	 */
 	useEffect(() => {
-		setScrollData({ isScrollable: false })
+		setScrollData(getScrollData(scrollRef.current))
 	}, [windowSize])
-
-	const { isScrollable, isAtStart, isAtEnd } = scrollData
 
 	return (
 		<div
@@ -63,7 +44,11 @@ function SidecarScrollContainer({ children }: { children: ReactNode }) {
 				[s.showBottomScrim]: isScrollable && !isAtEnd,
 			})}
 		>
-			<div className={s.scrollContainer} onScroll={handleScroll}>
+			<div
+				ref={scrollRef}
+				className={s.scrollContainer}
+				onScroll={handleScroll}
+			>
 				{children}
 			</div>
 		</div>
